refactor: move remove todo logic to todo-list.component
