+++ src/app/todo/components/todo-list/todo-list.component.html
@@ -16,6 +16,6 @@
       class="todo_item_remove"
       fontIcon="delete"
       aria-label="Remove todo"
-      (click)="onRemove(todo)" />
+      (click)="removeTodo(todo)" />
   </div>
 </div>
