+++ src/app/todo/components/todo-list/todo-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { Todo } from '../../../models';
+import { TodosService } from '../../../services/todos.service';
+import { ConfirmationModalService } from '../../../services/confirmation-modal.service';
 
 @Component({
   selector: 'app-todo-list',
@@ -10,13 +12,35 @@ export class TodoListComponent {
   @Input() todos: Todo[] = [];
 
   @Output() toggle = new EventEmitter<Todo>();
-  @Output() remove = new EventEmitter<Todo>();
+  @Output() refresh = new EventEmitter();
+
+  constructor(
+    public todosService: TodosService,
+    public modal: ConfirmationModalService
+  ) {}
 
   onToggle(todo: Todo) {
     this.toggle.emit(todo);
   }
 
-  onRemove(todo: Todo) {
-    this.remove.emit(todo);
+  removeTodo(todo: Todo) {
+    this.openModalThenRemove(todo);
+  }
+
+  private openModalThenRemove(todo: Todo) {
+    this.modal
+      .open()
+      .afterClosed()
+      .subscribe(confirm => {
+        if (confirm) {
+          this.removeThenReload(todo);
+        }
+      });
+  }
+
+  private removeThenReload(todo: Todo) {
+    this.todosService.removeTodo(todo).subscribe(() => {
+      this.refresh.emit();
+    });
   }
 }
